pyproject.toml
[tool.poetry]
name = "fastapi-django-example"
version = "0.1.0"
description = ""
authors = ["albertalexandrov <albert-ugatu@list.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
uvicorn = "^0.35.0"
orjson = "^3.10.18"
asyncpg = "^0.30.0"
greenlet = "^3.2.2"
prometheus-fastapi-instrumentator = "^7.1.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.28.0"}
fastapi-filter = "^2.0.1"
pydantic-i18n = "^0.4.5"
python-benedict = "^0.34.1"
faststream = {extras = ["redis"], version = "^0.5.41"}
fastapi-django = {path = "/Users/albertaleksandrov/PycharmProjects/fastapi-django"}
environs = "^14.2.0"
pytest = "^8.4.1"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
cryptography = "^45.0.5"
pydantic-async-validation = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
